Update README.md
The compound score is found by finding the sum of the valence scores, which are each between -4 and 4, and then normalizing it using an equation in the form x/sqrt(x^2+a). When x reaches really high or really low values, this score approaches 1 or -1 respectively. But it behaves in a less stable manner for lower values of x. This wouldn't exactly work well for extremely large sets of data as the sum of the valence scores would reach such a large magnitude that even if the overall sentiment of the text is fairly neutral, the a wouldn't be able to handle the immense value for x and would show a value that approaches -1 or 1 depending on the skew of the data set. This wouldn't really apply to this specific data set as it isn't exactly that massive plus it was already incredibly skewed to one side to begin with.